pyproject.toml
[tool.poetry]
name = "pyhid-usb-relay"
version = "0.1.0"
description = "A tool for controlling USB HID relays"
authors = ["Joshua Watt <JPEWhacker@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/JPEWdev/pyhid-usb-relay"
repository = "https://github.com/JPEWdev/pyhid-usb-relay"
classifiers = [
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.6"
xdg = "^5.1.1"
pyusb = "^1.2.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
pyhid-usb-relay = "pyhid_usb_relay.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
